requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv==1.0.1

# Machine Learning
scikit-learn>=1.4.0
pandas>=2.2.0
numpy
scipy>=1.12.0
joblib>=1.3.0
lightgbm>=4.1.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx==0.25.1
docker>=6.1.3

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Security
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Added from the code block
requests>=2.26.0

# New dependencies
pytesseract==0.3.10
Pillow>=10.2.0
opencv-python-headless==4.8.1.78
aiohttp>=3.8.1
python-dateutil>=2.8.2
shap